asdf

+++ mkusers.sh
@@ -12,9 +12,11 @@ docker cp -a $CONT:/etc/skel ./skel
 for user in `cut -d: -f1 $USERS`; do
   echo $user
   HM=`grep ^$user $USERS | cut -d: -f6`
-  docker exec $CONT rm -rf $HM
-  docker cp -a ./skel $CONT:$HM
-  docker exec $CONT chown -R $user:`grep ^$user $USERS | cut -d: -f4` $HM
+  if ! docker exec $CONT test -d $HM; then
+    docker exec $CONT rm -rf $HM
+    docker cp -a ./skel $CONT:$HM
+    docker exec $CONT chown -R $user:`grep ^$user $USERS | cut -d: -f4` $HM
+  fi
 done
 rm -rf ./skel
 echo 'All done!'
